# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "keep track of your time",
    "version" : "1.0.0",
    "title" : "Time Keeping"
  },
  "schemes" : [ "https", "http" ],
  "paths" : {
...
    "/api/v1/tasks/{task_id}" : {
      "get" : {
        "tags" : [ "Task" ],
        "summary" : "Returns the Task with the id if it exists",
        "description" : "",
        "operationId" : "taskController.find",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "task_id",
          "in" : "path",
          "description" : "The id of the coresponding Task",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Task"
              }
            }
          },
          "400" : {
            "description" : "Invalid status value"
          }
        },
        "security" : [ {
          "taskstore_auth" : [ "write:tasks", "read:tasks" ]
        } ]
      },
      "put" : {
        "tags" : [ "Task" ],
        "summary" : "Updates an existing Task",
        "description" : "",
        "operationId" : "taskController.update",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "task_id",
          "in" : "path",
          "description" : "The id of the coresponding Task",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Task object that needs to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Task"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Task not found"
          },
          "405" : {
            "description" : "Validation exception"
          }
        },
        "security" : [ {
          "taskstore_auth" : [ "write:tasks", "read:tasks" ]
        } ]
      },
      "delete" : {
        "tags" : [ "Task" ],
        "summary" : "Deletes a Task",
        "description" : "",
        "operationId" : "taskController.destroy",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "api_key",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "task_id",
          "in" : "path",
          "description" : "The id of the coresponding task",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted task"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Task not found"
          }
        },
        "security" : [ {
          "taskstore_auth" : [ "write:tasks", "read:tasks" ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "taskstore_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "http://Taskstore.swagger.io/oauth/dialog",
      "flow" : "implicit",
      "scopes" : {
        "write:tasks" : "modify tasks",
        "read:tasks" : "read tasks",
        "write:users" : "modify users",
        "read:users" : "read users"
      }
    },
    "api_key" : {
      "type" : "apiKey",
      "name" : "api_key",
      "in" : "header"
    }
  },
  "definitions" : {
    "Task" : {
      "type" : "object",
      "required" : [ "content", "title" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "title" : {
          "type" : "string",
          "example" : "what to do"
        },
        "content" : {
          "type" : "string",
          "example" : "do the dishes"
        },
        "createdBy" : {
          "type" : "integer",
          "format" : "int64"
        },
        "onGoing" : {
          "type" : "boolean",
          "example" : "true"
        },
        "duration" : {
          "type" : "string",
          "example" : "2 hours"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "completed", "pending", "new", "abandoned" ]
        }
      },
      "xml" : {
        "name" : "Task"
      }
    },
...
  }
}